change jest config
