--- jest.config.ts
@@ -1,30 +1,0 @@
-import type { Config } from 'jest';
-import nextJest from 'next/jest.js';
-
-const createJestConfig = nextJest({
-  // Provide the path to your Next.js app to load next.config.js and .env files in your test environment
-  dir: './',
-});
-
-// Add any custom config to be passed to Jest
-const config: Config = {
-  coverageProvider: 'v8',
-  testEnvironment: 'jsdom',
-  setupFilesAfterEnv: ['<rootDir>/jest.setup.ts'],
-  moduleNameMapper: {
-    'next/image': '<rootDir>/__mocks__/nextImageMock.tsx',
-  },
-  collectCoverage: true,
-  collectCoverageFrom: [
-    'pages/**/*.tsx',
-    'pages/**/*.ts',
-    'components/**/*.tsx',
-    'components/**/*.ts',
-    '!**/node_modules/**',
-  ],
-  // Add more setup options before each test is run
-  // setupFilesAfterEnv: ['<rootDir>/jest.setup.ts'],
-};
-
-// createJestConfig is exported this way to ensure that next/jest can load the Next.js config which is async
-export default createJestConfig(config);
